Extract nasty question logic to hook
